# api/openapi.yaml
openapi: 3.1.0
info:
  title: Puerta Abierta API
  version: 1.0.0
  summary: STR operations API for Paraguay-focused owners and operators
  description: |
    REST API contract for multi-tenant short-term rental operations:
    organizations, properties, reservations, tasks, finance, and reporting.
servers:
  - url: https://api.puertaabierta.app/v1
    description: Production
  - url: http://localhost:3000/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /guests:
    get:
      tags: [Guests]
      summary: List guests
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: q
          in: query
          description: Search by name/email/phone
          schema:
            type: string
      responses:
        "200":
          description: Guest list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Guest"
                required: [data]
    post:
      tags: [Guests]
      summary: Create guest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGuestInput"
      responses:
        "201":
          description: Guest created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
  /guests/{guest_id}:
    get:
      tags: [Guests]
      summary: Get guest by id
      parameters:
        - $ref: "#/components/parameters/GuestIdPath"
      responses:
        "200":
          description: Guest detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Guests]
      summary: Update guest
      parameters:
        - $ref: "#/components/parameters/GuestIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGuestInput"
      responses:
        "200":
          description: Guest updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Guest"
...
  /message-templates:
    get:
      tags: [Messaging]
      summary: List message templates
      parameters:
        - $ref: "#/components/parameters/OrgIdQuery"
        - name: channel
          in: query
          schema:
            $ref: "#/components/schemas/MessageChannel"
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/MessageTemplate"
                required: [data]
    post:
      tags: [Messaging]
      summary: Create message template
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageTemplateInput"
      responses:
        "201":
          description: Template created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageTemplate"
  /messages/send:
    post:
      tags: [Messaging]
      summary: Queue outbound message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "202":
          description: Message queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageLog"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrgIdQuery:
      name: org_id
      in: query
      required: true
      schema:
        type: string
        format: uuid
...
    GuestIdPath:
      name: guest_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    UUID:
      type: string
      format: uuid
...
    MessageChannel:
      type: string
      enum: [whatsapp, email, sms]
    MessageStatus:
      type: string
      enum: [queued, sent, failed]
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
      required: [error, message]
...
    Guest:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        full_name:
          type: string
        email:
          type: string
          format: email
          nullable: true
        phone_e164:
          type: string
          nullable: true
      required: [id, organization_id, full_name]
...
    MessageTemplate:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        template_key:
          type: string
        name:
          type: string
        channel:
          $ref: "#/components/schemas/MessageChannel"
        language_code:
          type: string
        subject:
          type: string
          nullable: true
        body:
          type: string
        variables:
          type: array
          items:
            type: string
        is_active:
          type: boolean
      required:
        [id, organization_id, template_key, name, channel, language_code, body, variables, is_active]
    MessageLog:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        organization_id:
          $ref: "#/components/schemas/UUID"
        reservation_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        guest_id:
          $ref: "#/components/schemas/UUID"
          nullable: true
        channel:
          $ref: "#/components/schemas/MessageChannel"
        recipient:
          type: string
        status:
          $ref: "#/components/schemas/MessageStatus"
        scheduled_at:
          type: string
          format: date-time
          nullable: true
      required: [id, organization_id, channel, recipient, status]
...
    CreateGuestInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        full_name:
          type: string
        email:
          type: string
          format: email
        phone_e164:
          type: string
      required: [organization_id, full_name]
    UpdateGuestInput:
      type: object
      properties:
        full_name:
          type: string
        email:
          type: string
          format: email
        phone_e164:
          type: string
        notes:
          type: string
...
    CreateMessageTemplateInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        template_key:
          type: string
        name:
          type: string
        channel:
          $ref: "#/components/schemas/MessageChannel"
        language_code:
          type: string
        subject:
          type: string
        body:
          type: string
        variables:
          type: array
          items:
            type: string
      required: [organization_id, template_key, name, channel, body]
    SendMessageInput:
      type: object
      properties:
        organization_id:
          $ref: "#/components/schemas/UUID"
        reservation_id:
          $ref: "#/components/schemas/UUID"
        guest_id:
          $ref: "#/components/schemas/UUID"
        template_id:
          $ref: "#/components/schemas/UUID"
        channel:
          $ref: "#/components/schemas/MessageChannel"
        recipient:
          type: string
        variables:
          type: object
          additionalProperties: true
        scheduled_at:
          type: string
          format: date-time
      required: [organization_id, channel, recipient]
...